upgraded contract

+++ contract/main.sol
@@ -13,10 +13,7 @@ contract owned {
     // function is executed and otherwise, an exception is
     // thrown.
     modifier onlyOwner {
-        require(
-            msg.sender == owner,
-            "Only owner can call this function."
-        );
+        require(msg.sender == owner, "Only owner can call this function.");
         _;
     }
 }
@@ -33,6 +30,7 @@ contract mortal is owned {
 
 contract GetLogin is mortal {
     event EventStoreWallet(bytes32 indexed username, address indexed walletAddress, string ciphertext, string iv, string salt, string mac);
+    event EventInviteCreated(bytes32 indexed creatorUsername, address inviteAddress);
 
     uint8 sessionMain = 1;
     uint8 sessionApp = 2;
@@ -40,11 +38,13 @@ contract GetLogin is mortal {
     struct Username
     {
         bool isActive;
+        // todo define a uniform variable name
         bytes32 username;
     }
 
     struct UserInfo
     {
+        // todo define a uniform variable name
         bytes32 username;
         bool isActive;
     }
@@ -52,24 +52,30 @@ contract GetLogin is mortal {
     struct InviteInfo
     {
         address inviteAddress;
+        // todo define a uniform variable name
         bytes32 creatorUsername;
+        bytes32 registeredUsername;
         bool isActive;
     }
 
     struct UserSession
     {
+        // todo define a uniform variable name
         bytes32 username;
         address wallet;
         uint8 sessionType;
+        uint64 appId;
     }
 
     struct Application
     {
         uint64 id;
+        // todo define a uniform variable name
         bytes32 usernameHash;
         string title;
         string description;
         string[] allowedUrls;
+        bool isActive;
     }
 
     uint64 public applicationId = 1;
@@ -81,18 +87,7 @@ contract GetLogin is mortal {
     mapping(uint64 => Application) public Applications;
 
     modifier userRegistered {
-        require(
-            isAddressRegistered(msg.sender),
-            "Only registered user can call this function"
-        );
-        _;
-    }
-
-    modifier userInvited {
-        require(
-            msg.sender == owner,
-            "Only invited user can call this function"
-        );
+        require(isAddressRegistered(msg.sender), "Only registered user can call this function");
         _;
     }
 
@@ -105,22 +100,19 @@ contract GetLogin is mortal {
 
     /* Private methods */
     function _createUser(bytes32 usernameHash) private {
-        // todo emit event
-        require(
-            isUsernameExists(usernameHash) == false,
-            "Username already used");
-        require(
-            isAddressRegistered(msg.sender) == false,
-            "Wallet already used");
+        // todo emit event?
+        require(isUsernameExists(usernameHash) == false, "Username already used");
+        require(isAddressRegistered(msg.sender) == false, "Wallet already used");
         Users[usernameHash] = UserInfo({username: usernameHash, isActive: true});
         UsersAddressUsername[msg.sender] = Username({username: usernameHash, isActive: true});
-        addSession(msg.sender, sessionMain);
+        addMainSession(msg.sender);
     }
 
     function _createApplication(bytes32 usernameHash, string memory title, string memory description) private returns (uint64) {
+        // todo only main session can create and edit app?
         // todo emit event?
         string[] memory allowedUrls;
-        Applications[applicationId] = Application({id: applicationId, usernameHash: usernameHash, title: title, description: description, allowedUrls: allowedUrls});
+        Applications[applicationId] = Application({id: applicationId, usernameHash: usernameHash, title: title, description: description, allowedUrls: allowedUrls, isActive: true});
         applicationId++;
 
         return applicationId;
@@ -136,21 +128,53 @@ contract GetLogin is mortal {
         delete Applications[appId].allowedUrls[index];
     }
 
+    function _addSession(address wallet, uint8 sessionType, uint64 appId) private {
+        validateAppExists(appId);
+        require(isAddressRegistered(wallet) == true, "Address already used");
+        bytes32 usernameHash = getUsernameByAddress(wallet);
+        UserSessions[usernameHash].push(UserSession({username: usernameHash, wallet: wallet, sessionType: sessionType, appId: appId}));
+    }
+
     /* End of private methods */
 
-    /* Public methods with userRegistered modifier */
+    /* Validators */
+    function validateAppOwner(uint64 appId, address wallet) public view userRegistered {
+        require(isAppOwner(appId, wallet) == true, "You do not have access to this application");
+    }
+
+    function validateAppExists(uint64 appId) public view userRegistered {
+        // todo return app info?
+        getApplication(appId);
+    }
+
+    function validateInviteActive(address wallet) public view userRegistered {
+        require(isActiveInvite(wallet) == true, "Invite not active");
+    }
+
+    /* End validators */
+
+
+    /* Public methods */
     function createApplication(string memory title, string memory description) public userRegistered {
         bytes32 usernameHash = getUsernameByAddress(msg.sender);
         _createApplication(usernameHash, title, description);
     }
 
     function addApplicationUrl(uint64 appId, string memory url) public userRegistered {
-        // todo check is owner
+        validateAppOwner(appId, msg.sender);
         _addApplicationUrl(appId, url);
     }
 
+    function renameApplication(uint64 appId, string memory title, string memory description) public userRegistered {
+        validateAppOwner(appId, msg.sender);
+        Application storage app = Applications[appId];
+        require(app.isActive == true, "App not found");
+        app.title = title;
+        app.description = description;
+    }
+
     function deleteApplicationUrl(uint64 appId, uint index) public userRegistered {
-        // todo check is owner
+        validateAppOwner(appId, msg.sender);
         _deleteApplicationUrl(appId, index);
     }
 
@@ -159,28 +183,40 @@ contract GetLogin is mortal {
     }
 
     function createInvite(address payable inviteAddress, bytes32 creatorUsername) public payable userRegistered {
-        // todo emit event?
-        Invites[inviteAddress] = InviteInfo({inviteAddress: inviteAddress, creatorUsername: creatorUsername, isActive: true});
+        Invites[inviteAddress] = InviteInfo({inviteAddress: inviteAddress, creatorUsername: creatorUsername, registeredUsername: '', isActive: true});
+        emit EventInviteCreated(creatorUsername, inviteAddress);
     }
 
     function createUserFromInvite(bytes32 usernameHash, address walletAddress, string memory ciphertext, string memory iv, string memory salt, string memory mac) public payable userRegistered {
-       // todo check is sender can register
-       // todo disable sender for reg
+       validateInviteActive(msg.sender);
+       require(isAddressRegistered(walletAddress) == false, "Address already registered");
+       InviteInfo storage invite = Invites[msg.sender];
        createUser(usernameHash);
        emit EventStoreWallet(usernameHash, walletAddress, ciphertext, iv, salt, mac);
+       invite.isActive = false;
+       invite.registeredUsername = usernameHash;
     }
 
-    function addSession(address wallet, uint8 sessionType) public payable userRegistered {
-        // todo check is username assigned with sender
-        // todo check is wallet not already exists
-        // todo check is wallet can add new wallets
-        bytes32 usernameHash = getUsernameByAddress(wallet);
-        UserSessions[usernameHash].push(UserSession({username: usernameHash, wallet: wallet, sessionType: sessionType}));
+    function addAppSession(address wallet, uint64 appId) public payable userRegistered {
+        // todo check only one main session possible
+        // todo hide user apps ids?
+        _addSession(wallet, sessionApp, appId);
     }
 
-    /* End of public methods with userRegistered modifier */
+    function addMainSession(address wallet) public payable userRegistered {
+        _addSession(wallet, sessionApp, 0);
+    }
+
+    /* End of public methods */
 
     /* View methods */
+    function getApplication(uint64 id) public view returns (Application memory) {
+        Application memory app = Applications[id];
+        require(app.isActive == true, "App not found");
+
+        return app;
+    }
+
     function getUserInfo(bytes32 usernameHash) public view returns (UserInfo memory) {
         return Users[usernameHash];
     }
@@ -194,21 +230,26 @@ contract GetLogin is mortal {
     }
 
     function isAppOwner(uint64 appIp, address checkAddress) public view returns (bool) {
-        // todo implemennt
-        //return getUserByAddress(wallet).isActive == true;
+        bytes32 currentUsernameHash = getUsernameByAddress(checkAddress);
+        return getApplication(appIp).usernameHash == currentUsernameHash;
     }
 
     function getUserByAddress(address wallet) public view returns (UserInfo memory) {
         Username memory currentUser = UsersAddressUsername[wallet];
-        require(
-            currentUser.isActive == true,
-            "User with this address not found"
-        );
+        require( currentUser.isActive == true, "User with this address not found");
         return Users[currentUser.username];
     }
 
     function getUsernameByAddress(address wallet) public view returns (bytes32) {
         return getUserByAddress(wallet).username;
+    }
+
+    function isActiveInvite(address wallet) public view returns (bool) {
+        return Invites[wallet].isActive == true;
+    }
+
+    function getInvite(address wallet) public view returns (InviteInfo memory) {
+        return Invites[wallet];
     }
 
     /* End of view methods */
